StartMicrovmError is exposed by the builder
Also remove unused StartMicrovmErrors
