+++ src/vmm/src/error.rs
@@ -7,12 +7,12 @@ use std::fmt::{Display, Formatter};
 use std::io;
 
 use super::{
-    device_manager, vmm_config::drive::DriveError, vmm_config::machine_config::VmConfigError,
-    vmm_config::net::NetworkInterfaceError, vstate,
+    vmm_config::drive::DriveError, vmm_config::machine_config::VmConfigError,
+    vmm_config::net::NetworkInterfaceError,
 };
-use kernel::loader as kernel_loader;
-use memory_model::GuestMemoryError;
+use device_manager;
 use polly::event_manager;
+use vstate;
 
 /// Errors associated with the VMM internal logic. These errors cannot be generated by direct user
 /// input, but can result from bad configuration of the host (for example if Firecracker doesn't
@@ -105,164 +105,6 @@ impl std::fmt::Display for Error {
     }
 }
 
-/// Errors associated with starting the instance.
-// TODO: add error kind to these variants because not all these errors are user or internal.
-#[derive(Debug)]
-pub enum StartMicrovmError {
-    /// Cannot configure the VM.
-    ConfigureVm(vstate::Error),
-    /// Unable to seek the block device backing file due to invalid permissions or
-    /// the file was deleted/corrupted.
-    CreateBlockDevice(std::io::Error),
-    /// Split this at some point.
-    /// Internal errors are due to resource exhaustion.
-    /// Users errors are due to invalid permissions.
-    CreateNetDevice(devices::virtio::Error),
-    /// Failed to create a `RateLimiter` object.
-    CreateRateLimiter(std::io::Error),
-    /// Failed to create the backend for the vsock device.
-    CreateVsockBackend(devices::virtio::vsock::VsockUnixBackendError),
-    /// Failed to create the vsock device.
-    CreateVsockDevice(devices::virtio::vsock::VsockError),
-    /// The device manager was not configured.
-    DeviceManager,
-    /// Cannot read from an Event file descriptor.
-    EventFd,
-    /// Memory regions are overlapping or mmap fails.
-    GuestMemory(GuestMemoryError),
-    // Temporarily added for mixing calls that may return an Error with others that may return a
-    // StartMicrovmError within the same function.
-    /// Internal error encountered while starting a microVM.
-    Internal(Error),
-    /// The kernel command line is invalid.
-    KernelCmdline(String),
-    /// Cannot load kernel due to invalid memory configuration or invalid kernel image.
-    KernelLoader(kernel_loader::Error),
-    /// Cannot load command line string.
-    LoadCommandline(kernel::cmdline::Error),
-    /// The start command was issued more than once.
-    MicroVMAlreadyRunning,
-    /// Cannot start the VM because the kernel was not configured.
-    MissingKernelConfig,
-    /// The net device configuration is missing the tap device.
-    NetDeviceNotConfigured,
-    /// Cannot open the block device backing file.
-    OpenBlockDevice(std::io::Error),
-    /// Cannot initialize a MMIO Block Device or add a device to the MMIO Bus.
-    RegisterBlockDevice(device_manager::mmio::Error),
-    /// Cannot initialize a MMIO Network Device or add a device to the MMIO Bus.
-    RegisterNetDevice(device_manager::mmio::Error),
-    /// Cannot initialize a MMIO Vsock Device or add a device to the MMIO Bus.
-    RegisterVsockDevice(device_manager::mmio::Error),
-    /// vCPUs were not configured.
-    VcpusNotConfigured,
-}
-
-/// It's convenient to automatically convert `kernel::cmdline::Error`s
-/// to `StartMicrovmError`s.
-impl std::convert::From<kernel::cmdline::Error> for StartMicrovmError {
-    fn from(e: kernel::cmdline::Error) -> StartMicrovmError {
-        StartMicrovmError::KernelCmdline(e.to_string())
-    }
-}
-
-impl Display for StartMicrovmError {
-    fn fmt(&self, f: &mut Formatter) -> std::fmt::Result {
-        use self::StartMicrovmError::*;
-        match *self {
-            ConfigureVm(ref err) => {
-                let mut err_msg = format!("{:?}", err);
-                err_msg = err_msg.replace("\"", "");
-
-                write!(f, "Cannot configure virtual machine. {}", err_msg)
-            }
-            CreateBlockDevice(ref err) => write!(
-                f,
-                "Unable to seek the block device backing file due to invalid permissions or \
-                 the file was deleted/corrupted. Error number: {}",
-                err
-            ),
-            CreateRateLimiter(ref err) => write!(f, "Cannot create RateLimiter: {}", err),
-            CreateVsockBackend(ref err) => {
-                write!(f, "Cannot create backend for vsock device: {:?}", err)
-            }
-            CreateVsockDevice(ref err) => write!(f, "Cannot create vsock device: {:?}", err),
-            CreateNetDevice(ref err) => {
-                let mut err_msg = format!("{:?}", err);
-                err_msg = err_msg.replace("\"", "");
-
-                write!(f, "Cannot create network device. {}", err_msg)
-            }
-            DeviceManager => write!(f, "The device manager was not configured."),
-            EventFd => write!(f, "Cannot read from an Event file descriptor."),
-            GuestMemory(ref err) => {
-                // Remove imbricated quotes from error message.
-                let mut err_msg = format!("{:?}", err);
-                err_msg = err_msg.replace("\"", "");
-                write!(f, "Invalid Memory Configuration: {}", err_msg)
-            }
-            Internal(ref err) => write!(f, "Internal error while starting microVM: {:?}", err),
-            KernelCmdline(ref err) => write!(f, "Invalid kernel command line: {}", err),
-            KernelLoader(ref err) => {
-                let mut err_msg = format!("{}", err);
-                err_msg = err_msg.replace("\"", "");
-                write!(
-                    f,
-                    "Cannot load kernel due to invalid memory configuration or invalid kernel \
-                     image. {}",
-                    err_msg
-                )
-            }
-            LoadCommandline(ref err) => {
-                let mut err_msg = format!("{}", err);
-                err_msg = err_msg.replace("\"", "");
-                write!(f, "Cannot load command line string. {}", err_msg)
-            }
-            MicroVMAlreadyRunning => write!(f, "Microvm already running."),
-            MissingKernelConfig => write!(f, "Cannot start microvm without kernel configuration."),
-            NetDeviceNotConfigured => {
-                write!(f, "The net device configuration is missing the tap device.")
-            }
-            OpenBlockDevice(ref err) => {
-                let mut err_msg = format!("{:?}", err);
-                err_msg = err_msg.replace("\"", "");
-
-                write!(f, "Cannot open the block device backing file. {}", err_msg)
-            }
-            RegisterBlockDevice(ref err) => {
-                let mut err_msg = format!("{}", err);
-                err_msg = err_msg.replace("\"", "");
-                write!(
-                    f,
-                    "Cannot initialize a MMIO Block Device or add a device to the MMIO Bus. {}",
-                    err_msg
-                )
-            }
-            RegisterNetDevice(ref err) => {
-                let mut err_msg = format!("{}", err);
-                err_msg = err_msg.replace("\"", "");
-
-                write!(
-                    f,
-                    "Cannot initialize a MMIO Network Device or add a device to the MMIO Bus. {}",
-                    err_msg
-                )
-            }
-            RegisterVsockDevice(ref err) => {
-                let mut err_msg = format!("{}", err);
-                err_msg = err_msg.replace("\"", "");
-
-                write!(
-                    f,
-                    "Cannot initialize a MMIO Vsock Device or add a device to the MMIO Bus. {}",
-                    err_msg
-                )
-            }
-            VcpusNotConfigured => write!(f, "vCPUs were not configured."),
-        }
-    }
-}
-
 use controller::{ErrorKind, VmmActionError};
 
 // It's convenient to turn DriveErrors into VmmActionErrors directly.
@@ -332,44 +174,6 @@ impl std::convert::From<NetworkInterfaceError> for VmmActionError {
         };
 
         VmmActionError::NetworkConfig(kind, e)
-    }
-}
-
-// It's convenient to turn StartMicrovmErrors into VmmActionErrors directly.
-impl std::convert::From<StartMicrovmError> for VmmActionError {
-    fn from(e: StartMicrovmError) -> Self {
-        use error::StartMicrovmError::*;
-
-        let kind = match e {
-            // User errors.
-            CreateVsockBackend(_)
-            | CreateBlockDevice(_)
-            | CreateNetDevice(_)
-            | KernelCmdline(_)
-            | KernelLoader(_)
-            | MicroVMAlreadyRunning
-            | MissingKernelConfig
-            | NetDeviceNotConfigured
-            | OpenBlockDevice(_)
-            | VcpusNotConfigured => ErrorKind::User,
-            // Internal errors.
-            ConfigureVm(_)
-            | CreateRateLimiter(_)
-            | CreateVsockDevice(_)
-            | DeviceManager
-            | EventFd
-            | GuestMemory(_)
-            | Internal(_)
-            | RegisterBlockDevice(_)
-            | RegisterNetDevice(_)
-            | RegisterVsockDevice(_) => ErrorKind::Internal,
-            // The only user `LoadCommandline` error is `CommandLineOverflow`.
-            LoadCommandline(ref cle) => match cle {
-                kernel::cmdline::Error::CommandLineOverflow => ErrorKind::User,
-                _ => ErrorKind::Internal,
-            },
-        };
-        VmmActionError::StartMicrovm(kind, e)
     }
 }
 
@@ -592,11 +396,6 @@ mod tests {
             ErrorKind::Internal
         );
         assert_eq!(
-            error_kind(StartMicrovmError::DeviceManager),
-            ErrorKind::Internal
-        );
-        assert_eq!(error_kind(StartMicrovmError::EventFd), ErrorKind::Internal);
-        assert_eq!(
             error_kind(StartMicrovmError::GuestMemory(
                 memory_model::GuestMemoryError::NoMemoryRegions
             )),
@@ -670,10 +469,6 @@ mod tests {
                 device_manager::mmio::Error::IrqsExhausted
             )),
             ErrorKind::Internal
-        );
-        assert_eq!(
-            error_kind(StartMicrovmError::VcpusNotConfigured),
-            ErrorKind::User
         );
     }
 
@@ -823,13 +618,6 @@ mod tests {
                 )
             ),
             "NetworkConfig(User, DeviceIdNotFound)"
-        );
-        assert_eq!(
-            format!(
-                "{:?}",
-                VmmActionError::StartMicrovm(ErrorKind::User, StartMicrovmError::EventFd)
-            ),
-            "StartMicrovm(User, EventFd)"
         );
         assert_eq!(
             format!(
